audio_book_pipeline/requirements.txt
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
ctranslate2==4.4.0
faster-whisper==1.1.0
whisperx==3.3.1
ffmpeg-python
pydub
langchain-google-genai
python-dotenv
pyyaml
srt
difflib
pyannote-audio
httpx
aiometer
tqdm